Fixed so initialize container will re-attempt to import database if db engine is yet not up and running

+++ dockerfiles/ezphp/prepare_distribution_volume.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Let's try to connect to db for 2 minutes ( 24 * 5 sec intervalls )
+MAXTRY=24
+
 cd /var/www
 
 
@@ -12,10 +15,19 @@ function prevent_multiple_execuition
     touch /already_run.txt
 }
 
+
+# $1 is description
 function set_splash_screen
 {
-    mv /var/www/web/index.php /var/www/web/index.php.org
-    echo "<html><body>Initializing</body></html>" > /var/www/web/index.php
+    if [ ! -f /var/www/web/index.php.org ]; then
+        mv /var/www/web/index.php /var/www/web/index.php.org
+    fi
+    echo "<html><body>$1</body></html>" > /var/www/web/index.php
+}
+
+function remove_splash_screen
+{
+    mv /var/www/web/index.php.org /var/www/web/index.php
 }
 
 function set_permissions
@@ -35,9 +47,25 @@ function set_permissions
 
 function import_database
 {
-    # Wait 15 secs for mysql to boot up
-    sleep 15
-    mysql -u$MYSQL_USER -p$MYSQL_PASSWORD $MYSQL_DATABASE -h db< /dbdump/ezp.sql
+    local DBUP
+    local TRY
+    DBUP=false
+    TRY=1
+    while [ $DBUP == "false" ]; do
+        echo Contacting mysql, attempt :$TRY
+        set_splash_screen "Importing database"
+        mysql -u$MYSQL_USER -p$MYSQL_PASSWORD $MYSQL_DATABASE -h db< /dbdump/ezp.sql && DBUP="true"
+        if [ $DBUP == "false" ]; then
+            set_splash_screen "Attempt $TRY failed. Waiting for db connection"
+        fi
+        let TRY=$TRY+1
+        if [ $TRY -eq $MAXTRY ]; then
+            echo Max limit reached. Not able to connect to mysql
+            rm /already_run.txt
+            exit 1;
+        fi
+        sleep 5;
+    done
 }
 
 function warm_cache
@@ -45,15 +73,12 @@ function warm_cache
     php ezpublish/console cache:warmup --env=$EZ_ENVIRONMENT
 }
 
-function remove_splash_screen
-{
-    mv /var/www/web/index.php.org /var/www/web/index.php
-}
-
 prevent_multiple_execuition
-set_splash_screen
+set_splash_screen "Initializing"
 set_permissions
+set_splash_screen "Waiting for db connection"
 import_database
+set_splash_screen "Warming cache"
 warm_cache
 remove_splash_screen
